Time: 0 ms (100%), Space: 8.6 MB (65.1%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -5,7 +5,7 @@ public:
         long long rev = 0;
         while(x>0)
         {
-            int digit = x%10;
+            long long digit = x%10;
             rev = rev*10 + digit;
             x/=10;
